Move to separate file for clarity
