+++ providers/v1/secretserver/provider.go
@@ -40,8 +40,7 @@ var (
 	errInvalidSpec                   = errors.New("invalid specification for secret server provider")
 	errClusterStoreRequiresNamespace = errors.New("when using a ClusterSecretStore, namespaces must be explicitly set")
 	errMissingSecretName             = errors.New("must specify a secret name")
-
-	errMissingSecretKey = errors.New("must specify a secret key")
+	errMissingSecretKey              = errors.New("must specify a secret key")
 )
 
 // Provider struct that implements the ESO esv1.Provider.
@@ -149,8 +148,8 @@ func loadCredentials(ctx context.Context, store esv1.GenericStore, cfg *esv1.Sec
 
 func secretServerCredentialRefPolicy(store esv1.GenericStore) esutils.ValueOrRefPolicy[esmeta.SecretKeySelector] {
 	return esutils.ValueOrRefPolicy[esmeta.SecretKeySelector]{
-		Presence:           esutils.RequireValueOrRef,
-		ValidateRef:        validateSecretServerCredentialSecretRef(store),
+		Presence:    esutils.RequireValueOrRef,
+		ValidateRef: validateSecretServerCredentialSecretRef(store),
 	}
 }
 
@@ -167,6 +166,27 @@ func validateSecretServerCredentialSecretRef(store esv1.GenericStore) func(esmet
 		}
 		return nil
 	}
+}
+
+func validateStoreSecretRef(store esv1.GenericStore, ref *esv1.SecretServerProviderRef) error {
+	return esutils.ValidateValueOrRef(ref.Value, ref.SecretRef, secretServerCredentialRefPolicy(store))
+}
+
+func validateSecretRef(ref *esv1.SecretServerProviderRef) error {
+	return esutils.ValidateValueOrRef(ref.Value, ref.SecretRef, esutils.ValueOrRefPolicy[esmeta.SecretKeySelector]{
+		Presence:    esutils.RequireValueOrRef,
+		ValidateRef: validateSecretServerCredentialSecretRefNameAndKey,
+	})
+}
+
+func validateSecretServerCredentialSecretRefNameAndKey(ref esmeta.SecretKeySelector) error {
+	if ref.Name == "" {
+		return errMissingSecretName
+	}
+	if ref.Key == "" {
+		return errMissingSecretKey
+	}
+	return nil
 }
 
 func doesConfigDependOnNamespace(cfg *esv1.SecretServerProvider) bool {
@@ -214,10 +234,10 @@ func getConfig(store esv1.GenericStore) (*esv1.SecretServerProvider, error) {
 		return nil, errEmptyPassword
 	}
 
-	if err := esutils.ValidateValueOrRef(cfg.Username.Value, cfg.Username.SecretRef, secretServerCredentialRefPolicy(store)); err != nil {
+	if err := validateStoreSecretRef(store, cfg.Username); err != nil {
 		return nil, err
 	}
-	if err := esutils.ValidateValueOrRef(cfg.Password.Value, cfg.Password.SecretRef, secretServerCredentialRefPolicy(store)); err != nil {
+	if err := validateStoreSecretRef(store, cfg.Password); err != nil {
 		return nil, err
 	}
 	return cfg, nil
