+++ runtime/esutils/utils_test.go
@@ -1109,95 +1109,6 @@ func TestCompareStringAndByteSlices(t *testing.T) {
 	}
 }
 
-func TestValidateValueOrRef(t *testing.T) {
-	errRef := errors.New("ref")
-	ref := "ref"
-
-	tests := []struct {
-		name    string
-		value   string
-		ref     *string
-		policy  ValueOrRefPolicy[string]
-		wantErr error
-	}{
-		{
-			name:  "requires exactly one: value",
-			value: "value",
-			policy: ValueOrRefPolicy[string]{
-				Presence: RequireValueOrRef,
-			},
-		},
-		{
-			name: "requires exactly one: ref",
-			ref:  &ref,
-			policy: ValueOrRefPolicy[string]{
-				Presence: RequireValueOrRef,
-			},
-		},
-		{
-			name:  "requires exactly one: rejects both",
-			value: "value",
-			ref:   &ref,
-			policy: ValueOrRefPolicy[string]{
-				Presence: RequireValueOrRef,
-			},
-			wantErr: errors.New("cannot specify both value and reference"),
-		},
-		{
-			name: "requires exactly one: rejects missing both",
-			policy: ValueOrRefPolicy[string]{
-				Presence: RequireValueOrRef,
-			},
-			wantErr: errors.New("must specify either value or reference"),
-		},
-		{
-			name: "validates ref when present",
-			ref:  &ref,
-			policy: ValueOrRefPolicy[string]{
-				Presence: RequireValueOrRef,
-				ValidateRef: func(string) error {
-					return errRef
-				},
-			},
-			wantErr: errRef,
-		},
-		{
-			name:  "allows optional empty pair",
-			value: "",
-			ref:   nil,
-			policy: ValueOrRefPolicy[string]{
-				Presence: AllowValueOrRef,
-			},
-		},
-		{
-			name: "requires ref only",
-			ref:  &ref,
-			policy: ValueOrRefPolicy[string]{
-				Presence: RequireRefOnly,
-			},
-		},
-		{
-			name:  "requires value only",
-			value: "value",
-			policy: ValueOrRefPolicy[string]{
-				Presence: RequireValueOnly,
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := ValidateValueOrRef(tt.value, tt.ref, tt.policy)
-			if got == nil && tt.wantErr == nil {
-				return
-			}
-			if got == nil || tt.wantErr == nil || got.Error() != tt.wantErr.Error() {
-				t.Errorf("ValidateValueOrRef() got = %v, want = %v", got, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestValidateSecretSelector(t *testing.T) {
 	tests := []struct {
 		desc     string
